Removed PRI macro which fails lint checker

+++ src/app/clusters/electrical-grid-conditions-server/tests/TestElectricalGridConditionsCluster.cpp
@@ -15,7 +15,6 @@
  */
 
 #include <app/clusters/electrical-grid-conditions-server/electrical-grid-conditions-server.h>
-#include <inttypes.h>
 #include <lib/support/logging/CHIPLogging.h>
 #include <memory>
 #include <platform/CHIPDeviceLayer.h>
@@ -33,8 +32,8 @@ bool gMatterReportingCalled = false;
 // This replaces the implementation in src/app/reporting/reporting.cpp
 void MatterReportingAttributeChangeCallback(uint16_t endpoint, uint32_t clusterId, uint32_t attributeId)
 {
-    ChipLogProgress(Test, "Mocked MatterReportingAttributeChangeCallback(ep=%" PRIu16 ", cluster=%" PRIu32 ", attr=%" PRIu32 ")",
-                    endpoint, clusterId, attributeId);
+    ChipLogProgress(Test, "Mocked MatterReportingAttributeChangeCallback(ep=%u, cluster=%u, attr=%u)",
+                    static_cast<unsigned int>(endpoint), static_cast<unsigned int>(clusterId), static_cast<unsigned int>(attributeId));
     gMatterReportingCalled = true;
 }
